requirements.txt
global-hotkeys==0.0.4
pyserial==3.5
pywin32==300
